added my presentation from ISDC 2018
The purpose of this upload is to share some introductory material explaining why this repo exists. I will also upload a more concise file to the same end. These documents are more for the code-drafters rather than the code-users, and will not be incorporated into the Multiplanetary Building Code (MBC) itself.

This is my long-form presentation in which I argue:
1. that an MBC is beneficial to humanity expanding off-world,
2. that now is the right time to start developing the MBC, and
3. that there are enough stakeholders who stand to gain from the MBC to justify its existence.